Modifications lucas

+++ core/viewer.cpp
@@ -174,7 +174,7 @@ void Viewer::insertMatrices(const QMatrix4x4& P,const QMatrix4x4& V,const QMatri
     _program->setUniformValue("pvm",P*V*M);
     } else {
         GLfloat matrix[16];
-        const qreal* data = (P).transposed().data();
+        const float* data = (P).transposed().data();
         for (int i = 0; i < 16; ++i) {
             matrix[i] = data[i];
         }
@@ -193,7 +193,7 @@ void Viewer::insertMatrices(const QMatrix4x4& P,const QMatrix4x4& V,const QMatri
         glGetFloatv (GL_PROJECTION_MATRIX, m);
 //        debugDataGL(m);
 //        Debugger::promptOpenGLError();
-        const qreal* data2= (M*V).data();
+        const float* data2= (M*V).data();
         for (int i = 0; i < 16; ++i) {
             matrix[i] = data2[i];
         }
